메뉴 인터페이스 추가

+++ project/s3-lcd-lvgl-demo/main/ap/stress/lv_demo_stress.c
@@ -74,7 +74,7 @@ static void obj_test_task_cb(lv_timer_t * tmr)
 
                 if(mem_free_start == 0)  mem_free_start = mon.free_size;
 
-                LV_LOG_USER("mem leak since start: %d, frag: %3d %%",  mem_free_start - mon.free_size, mon.frag_pct);
+                LV_LOG_USER("mem leak since start: %d, frag: %3d %%", (int)(mem_free_start - mon.free_size), mon.frag_pct);
             }
             break;
         case 0:
